+++ Player.h
@@ -16,7 +16,7 @@ public:
 	void Render(Tmpl8::Surface* screen);
 	void Update(float deltaTime);
 	void Shoot(bool fire);
-	void Rotate(int x, int y);
+	void Rotate(float x, float y);
 	MoveablePlayer* GetMoveable();
 
 	Spawner* GetSpawner();
@@ -29,7 +29,7 @@ private:
 	MoveablePlayer* mover;
 	Collider* col;
 	Spawner* spawner;
-	
+
 };
